@@ -20,32 +20,30 @@ public class RNPushNotificationBootEventReceiver extends BroadcastReceiver {
     public void onReceive(Context context, Intent intent) {
         Log.i(LOG_TAG, "RNPushNotificationBootEventReceiver loading scheduled notifications");
 
-        if (intent.getAction().equals("android.intent.action.BOOT_COMPLETED")) {
-            SharedPreferences sharedPreferences = context.getSharedPreferences(RNPushNotificationHelper.PREFERENCES_KEY, Context.MODE_PRIVATE);
-            Set<String> ids = sharedPreferences.getAll().keySet();
+        SharedPreferences sharedPreferences = context.getSharedPreferences(RNPushNotificationHelper.PREFERENCES_KEY, Context.MODE_PRIVATE);
+        Set<String> ids = sharedPreferences.getAll().keySet();
 
-            Application applicationContext = (Application) context.getApplicationContext();
-            RNPushNotificationHelper rnPushNotificationHelper = new RNPushNotificationHelper(applicationContext);
+        Application applicationContext = (Application) context.getApplicationContext();
+        RNPushNotificationHelper rnPushNotificationHelper = new RNPushNotificationHelper(applicationContext);
 
-            for (String id : ids) {
-                try {
-                    String notificationAttributesJson = sharedPreferences.getString(id, null);
-                    if (notificationAttributesJson != null) {
-                        RNPushNotificationAttributes notificationAttributes = RNPushNotificationAttributes.fromJson(notificationAttributesJson);
+        for (String id : ids) {
+            try {
+                String notificationAttributesJson = sharedPreferences.getString(id, null);
+                if (notificationAttributesJson != null) {
+                    RNPushNotificationAttributes notificationAttributes = RNPushNotificationAttributes.fromJson(notificationAttributesJson);
 
-                        if (notificationAttributes.getFireDate() < System.currentTimeMillis()) {
-                            Log.i(LOG_TAG, "RNPushNotificationBootEventReceiver: Showing notification for " +
-                                    notificationAttributes.getId());
-                            rnPushNotificationHelper.sendToNotificationCentre(notificationAttributes.toBundle());
-                        } else {
-                            Log.i(LOG_TAG, "RNPushNotificationBootEventReceiver: Scheduling notification for " +
-                                    notificationAttributes.getId());
-                            rnPushNotificationHelper.sendNotificationScheduledCore(notificationAttributes.toBundle());
-                        }
+                    if (notificationAttributes.getFireDate() < System.currentTimeMillis()) {
+                        Log.i(LOG_TAG, "RNPushNotificationBootEventReceiver: Showing notification for " +
+                                notificationAttributes.getId());
+                        rnPushNotificationHelper.sendToNotificationCentre(notificationAttributes.toBundle());
+                    } else {
+                        Log.i(LOG_TAG, "RNPushNotificationBootEventReceiver: Scheduling notification for " +
+                                notificationAttributes.getId());
+                        rnPushNotificationHelper.sendNotificationScheduledCore(notificationAttributes.toBundle());
                     }
-                } catch (Exception e) {
-                    Log.e(LOG_TAG, "Problem with boot receiver loading notification " + id, e);
                 }
+            } catch (Exception e) {
+                Log.e(LOG_TAG, "Problem with boot receiver loading notification " + id, e);
             }
         }
     }
