@@ -1,6 +1,7 @@
 package com.dieam.reactnativepushnotification.modules;
 
-
+import android.app.ActivityManager;
+import android.app.ActivityManager.RunningAppProcessInfo;
 import android.app.AlarmManager;
 import android.app.Application;
 import android.app.Notification;
@@ -15,22 +16,33 @@ import android.content.res.Resources;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
 import android.graphics.Color;
+import android.media.AudioAttributes;
 import android.media.RingtoneManager;
 import android.net.Uri;
 import android.os.Build;
 import android.os.Bundle;
-import androidx.core.app.NotificationCompat;
+import android.service.notification.StatusBarNotification;
 import android.util.Log;
 
+import androidx.annotation.RequiresApi;
+import androidx.core.app.NotificationCompat;
+
+import com.facebook.react.bridge.Arguments;
+import com.facebook.react.bridge.ReadableArray;
 import com.facebook.react.bridge.ReadableMap;
+import com.facebook.react.bridge.WritableArray;
+import com.facebook.react.bridge.WritableMap;
 
 import org.json.JSONArray;
 import org.json.JSONException;
 
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Calendar;
 import java.util.Date;
 import java.util.GregorianCalendar;
+import java.util.List;
+import java.util.Map;
 
 import static com.dieam.reactnativepushnotification.modules.RNPushNotification.LOG_TAG;
 import static com.dieam.reactnativepushnotification.modules.RNPushNotificationAttributes.fromJson;
@@ -38,7 +50,6 @@ import static com.dieam.reactnativepushnotification.modules.RNPushNotificationAt
 public class RNPushNotificationHelper {
     public static final String PREFERENCES_KEY = "rn_push_notification";
     private static final long DEFAULT_VIBRATION = 300L;
-    private static final String NOTIFICATION_CHANNEL_ID = "rn-push-notification-channel-id";
 
     private Context context;
     private RNPushNotificationConfig config;
@@ -71,14 +82,42 @@ public class RNPushNotificationHelper {
         return (AlarmManager) context.getSystemService(Context.ALARM_SERVICE);
     }
 
+    public void invokeApp(Bundle bundle) {
+        String packageName = context.getPackageName();
+        Intent launchIntent = context.getPackageManager().getLaunchIntentForPackage(packageName);
+        String className = launchIntent.getComponent().getClassName();
+
+        try {
+            Class<?> activityClass = Class.forName(className);
+            Intent activityIntent = new Intent(context, activityClass);
+
+            if(bundle != null) {
+                activityIntent.putExtra("notification", bundle);
+            }
+
+            activityIntent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+
+            context.startActivity(activityIntent);
+        } catch(Exception e) {
+            Log.e(LOG_TAG, "Class not found", e);
+            return;
+        }
+    }
+
     private PendingIntent toScheduleNotificationIntent(Bundle bundle) {
-        int notificationID = Integer.parseInt(bundle.getString("id"));
+        try {
+            int notificationID = Integer.parseInt(bundle.getString("id"));
 
-        Intent notificationIntent = new Intent(context, RNPushNotificationPublisher.class);
-        notificationIntent.putExtra(RNPushNotificationPublisher.NOTIFICATION_ID, notificationID);
-        notificationIntent.putExtras(bundle);
+            Intent notificationIntent = new Intent(context, RNPushNotificationPublisher.class);
+            notificationIntent.putExtra(RNPushNotificationPublisher.NOTIFICATION_ID, notificationID);
+            notificationIntent.putExtras(bundle);
 
-        return PendingIntent.getBroadcast(context, notificationID, notificationIntent, PendingIntent.FLAG_UPDATE_CURRENT);
+            return PendingIntent.getBroadcast(context, notificationID, notificationIntent, PendingIntent.FLAG_UPDATE_CURRENT);
+        } catch (Exception e) {
+            Log.e(LOG_TAG, "Unable to parse Notification ID", e);
+        }
+
+        return null;
     }
 
     public void sendNotificationScheduled(Bundle bundle) {
@@ -111,7 +150,7 @@ public class RNPushNotificationHelper {
 
         SharedPreferences.Editor editor = scheduledNotificationsPersistence.edit();
         editor.putString(id, notificationAttributes.toJson().toString());
-        commit(editor);
+        editor.apply();
 
         boolean isSaved = scheduledNotificationsPersistence.contains(id);
         if (!isSaved) {
@@ -123,21 +162,42 @@ public class RNPushNotificationHelper {
 
     public void sendNotificationScheduledCore(Bundle bundle) {
         long fireDate = (long) bundle.getDouble("fireDate");
+        boolean allowWhileIdle = bundle.getBoolean("allowWhileIdle");
 
         // If the fireDate is in past, this will fire immediately and show the
         // notification to the user
         PendingIntent pendingIntent = toScheduleNotificationIntent(bundle);
 
+        if (pendingIntent == null) {
+            return;
+        }
+
         Log.d(LOG_TAG, String.format("Setting a notification with id %s at time %s",
                 bundle.getString("id"), Long.toString(fireDate)));
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT) {
-            getAlarmManager().setExact(AlarmManager.RTC_WAKEUP, fireDate, pendingIntent);
+            if (allowWhileIdle && Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+                getAlarmManager().setExactAndAllowWhileIdle(AlarmManager.RTC_WAKEUP, fireDate, pendingIntent);
+            } else {
+                getAlarmManager().setExact(AlarmManager.RTC_WAKEUP, fireDate, pendingIntent);
+            }
         } else {
             getAlarmManager().set(AlarmManager.RTC_WAKEUP, fireDate, pendingIntent);
         }
     }
 
-    public void sendToNotificationCentre(Bundle bundle) {
+
+    public void sendToNotificationCentre(final Bundle bundle) {
+      RNPushNotificationPicturesAggregator aggregator = new RNPushNotificationPicturesAggregator(new RNPushNotificationPicturesAggregator.Callback() {
+        public void call(Bitmap largeIconImage, Bitmap bigPictureImage) {
+          sendToNotificationCentreWithPicture(bundle, largeIconImage, bigPictureImage);
+        }
+      });
+
+      aggregator.setLargeIconUrl(context, bundle.getString("largeIconUrl"));
+      aggregator.setBigPictureUrl(context, bundle.getString("bigPictureUrl"));
+    }
+
+    public void sendToNotificationCentreWithPicture(Bundle bundle, Bitmap largeIconBitmap, Bitmap bigPictureBitmap) {
         try {
             Class intentClass = getMainActivityClass();
             if (intentClass == null) {
@@ -147,7 +207,7 @@ public class RNPushNotificationHelper {
 
             if (bundle.getString("message") == null) {
                 // this happens when a 'data' notification is received - we do not synthesize a local notification in this case
-                Log.d(LOG_TAG, "Cannot send to notification centre because there is no 'message' field in: " + bundle);
+                Log.d(LOG_TAG, "Ignore this message if you sent data-only notification. Cannot send to notification centre because there is no 'message' field in: " + bundle);
                 return;
             }
 
@@ -170,7 +230,7 @@ public class RNPushNotificationHelper {
             final String priorityString = bundle.getString("priority");
 
             if (priorityString != null) {
-                switch(priorityString.toLowerCase()) {
+                switch (priorityString.toLowerCase()) {
                     case "max":
                         priority = NotificationCompat.PRIORITY_MAX;
                         break;
@@ -195,7 +255,7 @@ public class RNPushNotificationHelper {
             final String visibilityString = bundle.getString("visibility");
 
             if (visibilityString != null) {
-                switch(visibilityString.toLowerCase()) {
+                switch (visibilityString.toLowerCase()) {
                     case "private":
                         visibility = NotificationCompat.VISIBILITY_PRIVATE;
                         break;
@@ -209,43 +269,61 @@ public class RNPushNotificationHelper {
                         visibility = NotificationCompat.VISIBILITY_PRIVATE;
                 }
             }
+            
+            String channel_id = bundle.getString("channelId");
 
-            NotificationCompat.Builder notification = new NotificationCompat.Builder(context, NOTIFICATION_CHANNEL_ID)
+            if(channel_id == null) {
+                channel_id = this.config.getNotificationDefaultChannelId();
+            }
+            
+            NotificationCompat.Builder notification = new NotificationCompat.Builder(context, channel_id)
                     .setContentTitle(title)
                     .setTicker(bundle.getString("ticker"))
                     .setVisibility(visibility)
                     .setPriority(priority)
-                    .setAutoCancel(bundle.getBoolean("autoCancel", true));
+                    .setAutoCancel(bundle.getBoolean("autoCancel", true))
+                    .setOnlyAlertOnce(bundle.getBoolean("onlyAlertOnce", false));
+            
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) { // API 24 and higher
+                // Restore showing timestamp on Android 7+
+                // Source: https://developer.android.com/reference/android/app/Notification.Builder.html#setShowWhen(boolean)
+                boolean showWhen = bundle.getBoolean("showWhen", true);
 
-            String group = bundle.getString("group");
-            if (group != null) {
-                notification.setGroup(group);
+                notification.setShowWhen(showWhen);
             }
 
-            notification.setContentText(bundle.getString("message"));
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) { // API 26 and higher
+                // Changing Default mode of notification
+                notification.setDefaults(Notification.DEFAULT_LIGHTS);
+            }
+      
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT_WATCH) { // API 20 and higher
+              String group = bundle.getString("group");
 
-            String largeIcon = bundle.getString("largeIcon");
+              if (group != null) {
+                  notification.setGroup(group);
+              }
 
-            String subText = bundle.getString("subText");
-
-            if (subText != null) {
-                notification.setSubText(subText);
+              if (bundle.containsKey("groupSummary") || bundle.getBoolean("groupSummary")) {
+                  notification.setGroupSummary(bundle.getBoolean("groupSummary"));
+              }
             }
 
             String numberString = bundle.getString("number");
+
             if (numberString != null) {
                 notification.setNumber(Integer.parseInt(numberString));
             }
 
-            int smallIconResId;
-            int largeIconResId;
+            // Small icon
+            int smallIconResId = 0;
 
             String smallIcon = bundle.getString("smallIcon");
 
-            if (smallIcon != null) {
-                smallIconResId = res.getIdentifier(smallIcon, "mipmap", packageName);
-            } else {
-                smallIconResId = res.getIdentifier("ic_notification", "mipmap", packageName);
+            if (smallIcon != null && !smallIcon.isEmpty()) {
+              smallIconResId = res.getIdentifier(smallIcon, "mipmap", packageName);
+            } else if(smallIcon == null) {
+              smallIconResId = res.getIdentifier("ic_notification", "mipmap", packageName);
             }
 
             if (smallIconResId == 0) {
@@ -256,54 +334,88 @@ public class RNPushNotificationHelper {
                 }
             }
 
-            if (largeIcon != null) {
-                largeIconResId = res.getIdentifier(largeIcon, "mipmap", packageName);
-            } else {
-                largeIconResId = res.getIdentifier("ic_launcher", "mipmap", packageName);
-            }
-
-            Bitmap largeIconBitmap = BitmapFactory.decodeResource(res, largeIconResId);
-
-            if (largeIconResId != 0 && (largeIcon != null || Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP)) {
-                notification.setLargeIcon(largeIconBitmap);
-            }
-
             notification.setSmallIcon(smallIconResId);
+
+            // Large icon
+            if(largeIconBitmap == null) {
+                int largeIconResId = 0;
+
+                String largeIcon = bundle.getString("largeIcon");
+
+                if (largeIcon != null && !largeIcon.isEmpty()) {
+                  largeIconResId = res.getIdentifier(largeIcon, "mipmap", packageName);
+                } else if(largeIcon == null) {
+                  largeIconResId = res.getIdentifier("ic_launcher", "mipmap", packageName);
+                }
+
+                // Before Lolipop there was no large icon for notifications.
+                if (largeIconResId != 0 && (largeIcon != null || Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP)) {
+                    largeIconBitmap = BitmapFactory.decodeResource(res, largeIconResId);
+                }
+            }
+            
+            if (largeIconBitmap != null){
+              notification.setLargeIcon(largeIconBitmap);
+            }
+
+            String message = bundle.getString("message");
+
+            notification.setContentText(message);
+
+            String subText = bundle.getString("subText");
+
+            if (subText != null) {
+                notification.setSubText(subText);
+            }
+ 
             String bigText = bundle.getString("bigText");
 
             if (bigText == null) {
-                bigText = bundle.getString("message");
+                bigText = message;
             }
 
-            notification.setStyle(new NotificationCompat.BigTextStyle().bigText(bigText));
+            NotificationCompat.Style style;
+
+            if(bigPictureBitmap != null) {
+              style = new NotificationCompat.BigPictureStyle()
+                      .bigPicture(bigPictureBitmap)
+                      .setBigContentTitle(title)
+                      .setSummaryText(message);
+            }
+            else {
+              style = new NotificationCompat.BigTextStyle().bigText(bigText);
+            }
+
+            notification.setStyle(style);
 
             Intent intent = new Intent(context, intentClass);
             intent.addFlags(Intent.FLAG_ACTIVITY_SINGLE_TOP);
+            bundle.putBoolean("foreground", this.isApplicationInForeground());
             bundle.putBoolean("userInteraction", true);
             intent.putExtra("notification", bundle);
 
+            // Add message_id to intent so react-native-firebase/messaging can identify it
+            String messageId = bundle.getString("messageId");
+            if (messageId != null) {
+                intent.putExtra("message_id", messageId);
+            }
+
+            Uri soundUri = null;
+
             if (!bundle.containsKey("playSound") || bundle.getBoolean("playSound")) {
-                Uri soundUri = RingtoneManager.getDefaultUri(RingtoneManager.TYPE_NOTIFICATION);
                 String soundName = bundle.getString("soundName");
-                if (soundName != null) {
-                    if (!"default".equalsIgnoreCase(soundName)) {
 
-                        // sound name can be full filename, or just the resource name.
-                        // So the strings 'my_sound.mp3' AND 'my_sound' are accepted
-                        // The reason is to make the iOS and android javascript interfaces compatible
-
-                        int resId;
-                        if (context.getResources().getIdentifier(soundName, "raw", context.getPackageName()) != 0) {
-                            resId = context.getResources().getIdentifier(soundName, "raw", context.getPackageName());
-                        } else {
-                            soundName = soundName.substring(0, soundName.lastIndexOf('.'));
-                            resId = context.getResources().getIdentifier(soundName, "raw", context.getPackageName());
-                        }
-
-                        soundUri = Uri.parse("android.resource://" + context.getPackageName() + "/" + resId);
-                    }
+                if (soundName == null) {
+                    soundName = "default";
                 }
+
+                soundUri = getSoundUri(soundName);
+
                 notification.setSound(soundUri);
+            }
+
+            if (soundUri == null || Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+                notification.setSound(null);
             }
 
             if (bundle.containsKey("ongoing") || bundle.getBoolean("ongoing")) {
@@ -328,16 +440,44 @@ public class RNPushNotificationHelper {
                     PendingIntent.FLAG_UPDATE_CURRENT);
 
             NotificationManager notificationManager = notificationManager();
-            checkOrCreateChannel(notificationManager);
 
-            notification.setContentIntent(pendingIntent);
+            long[] vibratePattern = new long[]{0};
 
             if (!bundle.containsKey("vibrate") || bundle.getBoolean("vibrate")) {
                 long vibration = bundle.containsKey("vibration") ? (long) bundle.getDouble("vibration") : DEFAULT_VIBRATION;
                 if (vibration == 0)
                     vibration = DEFAULT_VIBRATION;
-                notification.setVibrate(new long[]{0, vibration});
+
+                vibratePattern = new long[]{0, vibration};
+
+                notification.setVibrate(vibratePattern); 
             }
+
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) { 
+              // Define the shortcutId
+              String shortcutId = bundle.getString("shortcutId");
+              
+              if (shortcutId != null) {
+                notification.setShortcutId(shortcutId);
+              }
+ 
+              Long timeoutAfter = (long) bundle.getDouble("timeoutAfter");
+  
+              if (timeoutAfter != null && timeoutAfter >= 0) {
+                notification.setTimeoutAfter(timeoutAfter);
+              }
+            }
+
+            Long when = (long) bundle.getDouble("when");
+  
+            if (when != null && when >= 0) {
+              notification.setWhen(when);
+            }
+
+            notification.setUsesChronometer(bundle.getBoolean("usesChronometer", false));
+                
+            notification.setChannelId(channel_id);
+            notification.setContentIntent(pendingIntent);
 
             JSONArray actionsArray = null;
             try {
@@ -360,16 +500,23 @@ public class RNPushNotificationHelper {
                         continue;
                     }
 
-                    Intent actionIntent = new Intent(context, intentClass);
+
+                    Intent actionIntent = new Intent(context, RNPushNotificationActions.class);
+                    actionIntent.setAction(packageName + ".ACTION_" + i);
+
                     actionIntent.addFlags(Intent.FLAG_ACTIVITY_SINGLE_TOP);
-                    actionIntent.setAction(context.getPackageName() + "." + action);
 
                     // Add "action" for later identifying which button gets pressed.
                     bundle.putString("action", action);
                     actionIntent.putExtra("notification", bundle);
+                    actionIntent.setPackage(packageName);
+                    if (messageId != null) {
+                        intent.putExtra("message_id", messageId);
+                    }
 
-                    PendingIntent pendingActionIntent = PendingIntent.getActivity(context, notificationID, actionIntent,
+                    PendingIntent pendingActionIntent = PendingIntent.getBroadcast(context, notificationID, actionIntent,
                             PendingIntent.FLAG_UPDATE_CURRENT);
+                    
                     if(action.equals("ReplyInput")){
                         //Action with inline reply
                         if(android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.KITKAT_WATCH){
@@ -390,8 +537,11 @@ public class RNPushNotificationHelper {
                         }
                     }
                     else{
-                        // Add "action" for later identifying which button gets pressed
-                        notification.addAction(icon, action, pendingActionIntent);
+                        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+                            notification.addAction(new NotificationCompat.Action.Builder(icon, action, pendingActionIntent).build());
+                        } else {
+                          notification.addAction(icon, action, pendingActionIntent);
+                        }
                     }
             }
 
@@ -406,17 +556,19 @@ public class RNPushNotificationHelper {
             if (scheduledNotificationsPersistence.getString(notificationIdString, null) != null) {
                 SharedPreferences.Editor editor = scheduledNotificationsPersistence.edit();
                 editor.remove(notificationIdString);
-                commit(editor);
+                editor.apply();
             }
 
-            Notification info = notification.build();
-            info.defaults |= Notification.DEFAULT_LIGHTS;
+            if (!(this.isApplicationInForeground() && bundle.getBoolean("ignoreInForeground"))) {
+                Notification info = notification.build();
+                info.defaults |= Notification.DEFAULT_LIGHTS;
 
-            if (bundle.containsKey("tag")) {
-                String tag = bundle.getString("tag");
-                notificationManager.notify(tag, notificationID, info);
-            } else {
-                notificationManager.notify(notificationID, info);
+                if (bundle.containsKey("tag")) {
+                    String tag = bundle.getString("tag");
+                    notificationManager.notify(tag, notificationID, info);
+                } else {
+                    notificationManager.notify(notificationID, info);
+                }
             }
 
             // Can't use setRepeating for recurring notifications because setRepeating
@@ -470,7 +622,7 @@ public class RNPushNotificationHelper {
                     nextEvent.set(Calendar.YEAR, nextEvent.get(Calendar.YEAR) + (nextMonth == 0 ? 1 : 0));
                     nextEvent.set(Calendar.MONTH, nextMonth);
                     final int maxDay = nextEvent.getActualMaximum(Calendar.DAY_OF_MONTH);
-                    nextEvent.set(Calendar.DAY_OF_MONTH, fireDay <= maxDay ? fireDay : maxDay);
+                    nextEvent.set(Calendar.DAY_OF_MONTH, Math.min(fireDay, maxDay));
                     nextEvent.set(Calendar.HOUR_OF_DAY, fireHour);
                     nextEvent.set(Calendar.MINUTE, fireMinute);
                     nextEvent.set(Calendar.SECOND, 0);
@@ -500,6 +652,27 @@ public class RNPushNotificationHelper {
         }
     }
 
+    private Uri getSoundUri(String soundName) {
+        if (soundName == null || "default".equalsIgnoreCase(soundName)) {
+            return RingtoneManager.getDefaultUri(RingtoneManager.TYPE_NOTIFICATION);
+        } else {
+
+            // sound name can be full filename, or just the resource name.
+            // So the strings 'my_sound.mp3' AND 'my_sound' are accepted
+            // The reason is to make the iOS and android javascript interfaces compatible
+
+            int resId;
+            if (context.getResources().getIdentifier(soundName, "raw", context.getPackageName()) != 0) {
+                resId = context.getResources().getIdentifier(soundName, "raw", context.getPackageName());
+            } else {
+                soundName = soundName.substring(0, soundName.lastIndexOf('.'));
+                resId = context.getResources().getIdentifier(soundName, "raw", context.getPackageName());
+            }
+
+            return Uri.parse("android.resource://" + context.getPackageName() + "/" + resId);
+        }
+    }
+
     public void clearNotifications() {
         Log.i(LOG_TAG, "Clearing alerts from the notification centre");
 
@@ -507,11 +680,78 @@ public class RNPushNotificationHelper {
         notificationManager.cancelAll();
     }
 
-    public void clearNotification(int notificationID) {
+    public void clearNotification(String tag, int notificationID) {
         Log.i(LOG_TAG, "Clearing notification: " + notificationID);
 
         NotificationManager notificationManager = notificationManager();
-        notificationManager.cancel(notificationID);
+        if(tag != null) {
+          notificationManager.cancel(tag, notificationID);
+        } else {
+          notificationManager.cancel(notificationID);
+        }
+    }
+
+    public void clearDeliveredNotifications(ReadableArray identifiers) {
+      NotificationManager notificationManager = notificationManager();
+      for (int index = 0; index < identifiers.size(); index++) {
+        String id = identifiers.getString(index);
+        Log.i(LOG_TAG, "Removing notification with id " + id);
+        notificationManager.cancel(Integer.parseInt(id));
+      }
+    }
+
+    @RequiresApi(api = Build.VERSION_CODES.M)
+    public WritableArray getDeliveredNotifications() {
+      NotificationManager notificationManager = notificationManager();
+      StatusBarNotification delivered[] = notificationManager.getActiveNotifications();
+      Log.i(LOG_TAG, "Found " + delivered.length + " delivered notifications");
+      WritableArray result = Arguments.createArray();
+      /*
+        * stay consistent to the return structure in
+        * https://facebook.github.io/react-native/docs/pushnotificationios.html#getdeliverednotifications
+        * but there is no such thing as a 'userInfo'
+        */
+      for (StatusBarNotification notification : delivered) {
+        Notification original = notification.getNotification();
+        Bundle extras = original.extras;
+        WritableMap notif = Arguments.createMap();
+        notif.putString("identifier", "" + notification.getId());
+        notif.putString("title", extras.getString(Notification.EXTRA_TITLE));
+        notif.putString("body", extras.getString(Notification.EXTRA_TEXT));
+        notif.putString("tag", notification.getTag());
+        notif.putString("group", original.getGroup());
+        result.pushMap(notif);
+      }
+
+      return result;
+
+    }
+
+    public WritableArray getScheduledLocalNotifications() {
+        WritableArray scheduled = Arguments.createArray();
+
+        Map<String, ?> scheduledNotifications = scheduledNotificationsPersistence.getAll();
+
+        for (Map.Entry<String, ?> entry : scheduledNotifications.entrySet()) {
+            try {
+                RNPushNotificationAttributes notification = fromJson(entry.getValue().toString());
+                WritableMap notificationMap = Arguments.createMap();
+
+                notificationMap.putString("title", notification.getTitle());
+                notificationMap.putString("message", notification.getMessage());
+                notificationMap.putString("number", notification.getNumber());
+                notificationMap.putDouble("date", notification.getFireDate());
+                notificationMap.putString("id", notification.getId());
+                notificationMap.putString("repeatInterval", notification.getRepeatType());
+                notificationMap.putString("soundName", notification.getSound());
+
+                scheduled.pushMap(notificationMap);
+            } catch (JSONException e) {
+                Log.e(LOG_TAG, e.getMessage());
+            }
+        }
+
+        return scheduled;
     }
 
     public void cancelAllScheduledNotifications() {
@@ -544,13 +784,17 @@ public class RNPushNotificationHelper {
         // remove it from the alarm manger schedule
         Bundle b = new Bundle();
         b.putString("id", notificationIDString);
-        getAlarmManager().cancel(toScheduleNotificationIntent(b));
+        PendingIntent pendingIntent = toScheduleNotificationIntent(b);
+
+        if (pendingIntent != null) {
+            getAlarmManager().cancel(pendingIntent);
+        }
 
         if (scheduledNotificationsPersistence.contains(notificationIDString)) {
             // remove it from local storage
             SharedPreferences.Editor editor = scheduledNotificationsPersistence.edit();
             editor.remove(notificationIDString);
-            commit(editor);
+            editor.apply();
         } else {
             Log.w(LOG_TAG, "Unable to find notification " + notificationIDString);
         }
@@ -558,70 +802,155 @@ public class RNPushNotificationHelper {
         // removed it from the notification center
         NotificationManager notificationManager = notificationManager();
 
-        notificationManager.cancel(Integer.parseInt(notificationIDString));
+        try {
+            notificationManager.cancel(Integer.parseInt(notificationIDString));
+        } catch (Exception e) {
+            Log.e(LOG_TAG, "Unable to parse Notification ID " + notificationIDString, e);
+        }
     }
 
     private NotificationManager notificationManager() {
         return (NotificationManager) context.getSystemService(Context.NOTIFICATION_SERVICE);
     }
 
-    private static void commit(SharedPreferences.Editor editor) {
-        if (Build.VERSION.SDK_INT < 9) {
-            editor.commit();
-        } else {
-            editor.apply();
-        }
+    public List<String> listChannels() {
+      List<String> channels = new ArrayList<>();
+      
+      if (Build.VERSION.SDK_INT < Build.VERSION_CODES.O)
+          return channels;
+      
+      NotificationManager manager = notificationManager();
+
+      if (manager == null)
+        return channels;
+
+      List<NotificationChannel> listChannels = manager.getNotificationChannels();
+
+      for(NotificationChannel channel : listChannels) {
+        channels.add(channel.getId());
+      }
+
+      return channels;
     }
 
-    private static boolean channelCreated = false;
-    private void checkOrCreateChannel(NotificationManager manager) {
+    public boolean channelBlocked(String channel_id) {
+      if (Build.VERSION.SDK_INT < Build.VERSION_CODES.O)
+          return false;
+      
+      NotificationManager manager = notificationManager();
+
+      if (manager == null)
+          return false;
+
+      NotificationChannel channel = manager.getNotificationChannel(channel_id);
+
+      if(channel == null)
+          return false;
+
+      return NotificationManager.IMPORTANCE_NONE == channel.getImportance();
+    }
+
+    public boolean channelExists(String channel_id) {
+      if (Build.VERSION.SDK_INT < Build.VERSION_CODES.O)
+          return false;
+      
+      NotificationManager manager = notificationManager();
+
+      if (manager == null)
+          return false;
+
+      NotificationChannel channel = manager.getNotificationChannel(channel_id);
+
+      return channel != null;
+    }
+
+    public void deleteChannel(String channel_id) {
         if (Build.VERSION.SDK_INT < Build.VERSION_CODES.O)
             return;
-        if (channelCreated)
-            return;
+        
+        NotificationManager manager = notificationManager();
+
         if (manager == null)
             return;
 
-        Bundle bundle = new Bundle();
+        manager.deleteNotificationChannel(channel_id);
+    }
 
-        int importance = NotificationManager.IMPORTANCE_HIGH;
-        final String importanceString = bundle.getString("importance");
+    private boolean checkOrCreateChannel(NotificationManager manager, String channel_id, String channel_name, String channel_description, Uri soundUri, int importance, long[] vibratePattern) {
+        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.O)
+            return false;
+        if (manager == null)
+            return false;
 
-        if (importanceString != null) {
-            switch(importanceString.toLowerCase()) {
-                case "default":
-                    importance = NotificationManager.IMPORTANCE_DEFAULT;
-                    break;
-                case "max":
-                    importance = NotificationManager.IMPORTANCE_MAX;
-                    break;
-                case "high":
-                    importance = NotificationManager.IMPORTANCE_HIGH;
-                    break;
-                case "low":
-                    importance = NotificationManager.IMPORTANCE_LOW;
-                    break;
-                case "min":
-                    importance = NotificationManager.IMPORTANCE_MIN;
-                    break;
-                case "none":
-                    importance = NotificationManager.IMPORTANCE_NONE;
-                    break;
-                case "unspecified":
-                    importance = NotificationManager.IMPORTANCE_UNSPECIFIED;
-                    break;
-                default:
-                    importance = NotificationManager.IMPORTANCE_HIGH;
+        NotificationChannel channel = manager.getNotificationChannel(channel_id);
+
+        if (
+          channel == null && channel_name != null && channel_description != null ||
+          channel != null &&
+          (
+            channel_name != null && !channel.getName().equals(channel_name) ||
+            channel_description != null && !channel.getDescription().equals(channel_description)
+          )
+        ) {
+            // If channel doesn't exist create a new one.
+            // If channel name or description is updated then update the existing channel.
+            channel = new NotificationChannel(channel_id, channel_name, importance);
+
+            channel.setDescription(channel_description);
+            channel.enableLights(true);
+            channel.enableVibration(vibratePattern != null);
+            channel.setVibrationPattern(vibratePattern);
+
+            if (soundUri != null) {
+                AudioAttributes audioAttributes = new AudioAttributes.Builder()
+                        .setContentType(AudioAttributes.CONTENT_TYPE_SONIFICATION)
+                        .setUsage(AudioAttributes.USAGE_NOTIFICATION)
+                        .build();
+
+                channel.setSound(soundUri, audioAttributes);
+            } else {
+                channel.setSound(null, null);
             }
+
+            manager.createNotificationChannel(channel);
+
+            return true;
         }
 
-        NotificationChannel channel = new NotificationChannel(NOTIFICATION_CHANNEL_ID, this.config.getChannelName() != null ? this.config.getChannelName() : "rn-push-notification-channel", importance);
+        return false;
+    }
 
-        channel.setDescription(this.config.getChannelDescription());
-        channel.enableLights(true);
-        channel.enableVibration(true);
+    public boolean createChannel(ReadableMap channelInfo) {
+        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.O)
+            return false;
 
-        manager.createNotificationChannel(channel);
-        channelCreated = true;
+        String channelId = channelInfo.getString("channelId");
+        String channelName = channelInfo.getString("channelName");
+        String channelDescription = channelInfo.hasKey("channelDescription") ? channelInfo.getString("channelDescription") : "";
+        String soundName = channelInfo.hasKey("soundName") ? channelInfo.getString("soundName") : "default";
+        int importance = channelInfo.hasKey("importance") ? channelInfo.getInt("importance") : 4;
+        boolean vibrate = channelInfo.hasKey("vibrate") && channelInfo.getBoolean("vibrate");
+        long[] vibratePattern = vibrate ? new long[] { 0, DEFAULT_VIBRATION } : null;
+
+        NotificationManager manager = notificationManager();
+
+        Uri soundUri = getSoundUri(soundName);
+
+        return checkOrCreateChannel(manager, channelId, channelName, channelDescription, soundUri, importance, vibratePattern);
+    }
+    
+    public boolean isApplicationInForeground() {
+        ActivityManager activityManager = (ActivityManager) context.getSystemService(Context.ACTIVITY_SERVICE);
+        List<RunningAppProcessInfo> processInfos = activityManager.getRunningAppProcesses();
+        if (processInfos != null) {
+            for (RunningAppProcessInfo processInfo : processInfos) {
+                if (processInfo.processName.equals(context.getPackageName())
+                        && processInfo.importance == RunningAppProcessInfo.IMPORTANCE_FOREGROUND
+                        && processInfo.pkgList.length > 0) {
+                    return true;
+                }
+            }
+        }
+        return false;
     }
 }
